refactor core: use a small_vector in http2 headers
fix
commit_hash:ba3c7099738ec2d09b8bae160a9221d0e4063c30

+++ core/src/server/net/connection.cpp
@@ -349,7 +349,8 @@ void Connection::SendResponse(request::RequestBase& request) {
           // NOLINTNEXTLINE(cppcoreguidelines-pro-type-static-cast-downcast)
           auto http2_session = static_cast<http::Http2Session*>(parser.get());
           http2_session->UpgradeToHttp2(h);
-          http_response.SetStreamId(http::kStreamIdAfterUpgradeResponse);
+          http_response.SetStreamId(
+              static_cast<std::int32_t>(http::kStreamIdAfterUpgradeResponse));
 
           http::WriteHttp2ResponseToSocket(http_response, *http2_session);
           parser_ = std::move(parser);
